requirements.txt
alive_progress
CurrencyConverter
pandas
python-steam-api
python-decouple
bs4
zope
